requirements.txt
Flask==3.0.0
deep-translator==1.11.4
langdetect==1.0.9
SpeechRecognition==3.10.0
pydub==0.25.1
Werkzeug==3.0.1
gunicorn==21.2.0
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sentry-sdk[flask]==1.40.0
Flask-Login==0.6.3
authlib==1.3.0
Flask-Mail==0.9.1
reportlab==4.0.9
python-docx==1.1.2
openai==1.40.0
httpx==0.27.2
gTTS==2.5.4

